thirteenth

+++ actions/workflows/deleteWorkflow.ts
@@ -1,20 +1,22 @@
 "use server";
 
 import prisma from "@/lib/prisma";
-import { auth } from "@clerk/nextjs/server";
+import { auth } from "@/auth";
 import { revalidatePath } from "next/cache";
 
 export async function DeleteWorkflow(id: string) {
-    const session = await auth();
-    const {userId} = session;
+  const session = await auth();
 
-    if (!userId) {
-        throw new Error("unathenticated")
-    }
-await prisma.workflow.delete({
+  if (!session?.user?.id) {
+    throw new Error("unathenticated");
+  }
+
+  const userId = session.user.id;
+
+  await prisma.workflow.delete({
     where: {
-        id,
-        userId,
+      id,
+      userId,
     },
   });
